pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name        = "contoso-fabric-data-platform"
version     = "1.4.0"
description = "Contoso enterprise data platform built on Microsoft Fabric"
readme      = "README.md"
license     = { file = "LICENSE" }
requires-python = ">=3.11"

authors = [
  { name = "Contoso Data Engineering", email = "data-engineering@contoso.example" }
]

keywords = ["fabric", "delta", "pyspark", "data-engineering", "azure"]

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "azure-identity>=1.15.0",
  "azure-keyvault-secrets>=4.7.0",
  "pyodbc>=4.0.39",
  "pandas>=2.1.0",
  "requests>=2.31.0",
  "tenacity>=8.2.3",
  "python-dotenv>=1.0.0",
  "pydantic>=2.5.0",
]

[project.optional-dependencies]
spark = [
  "pyspark>=3.4.0",
  "delta-spark>=2.4.0",
]
dev = [
  "pyspark>=3.4.0",
  "delta-spark>=2.4.0",
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "pytest-timeout>=2.2.0",
  "ruff>=0.1.0",
  "black>=23.0.0",
  "mypy>=1.5.0",
  "bandit[toml]>=1.7.5",
  "sqlfluff>=2.3.0",
  "ipykernel>=6.27.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
]
deploy = [
  "azure-identity>=1.15.0",
  "requests>=2.31.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

# ── Ruff (linting) ─────────────────────────────────────────────────────────────
[tool.ruff]
line-length    = 100
target-version = "py311"
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "UP",  # pyupgrade
]
ignore = [
  "E501",  # line too long (handled by black)
  "B008",  # do not perform function calls in default arguments
]

[tool.ruff.per-file-ignores]
"tests/**" = ["B", "F811"]

# ── Black (formatting) ─────────────────────────────────────────────────────────
[tool.black]
line-length    = 100
target-version = ["py311"]

# ── MyPy (type checking) ───────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
warn_return_any        = true
warn_unused_configs    = true
ignore_missing_imports = true

# ── Pytest ─────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths    = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts      = "--tb=short -q"
timeout      = 300

# ── Bandit (security scanning) ─────────────────────────────────────────────────
[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips        = ["B101"]  # assert usage in tests is fine

# ── Coverage ───────────────────────────────────────────────────────────────────
[tool.coverage.run]
source    = ["src"]
omit      = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
fail_under = 75
show_missing = true
